Fix the rolling period value to 144 + clear stored exposures
Setting 0 for rolling period, the value that is returned by default on a TemporaryExposureKey does not seem to work anymore when importing a key. Since a TEK in the gaen protocol currently always has a validity of 1 day, this value can be considered a constant of 144.
Clear the stored exposures before importing the key.